100-prime_factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
   */
 int main(void)
 {
-	long long n = 612852475143;
+	long n = 612852475143;
 	int i = 2;
 	int max = 0;
